src/providers: Implement DuckDNS & error module
Implement the DuckDNS (https://duckdns.org) as the first supported
dynamic DNS provider.

The commit also has adjusted some general types which were implemented
earlier to fit the first implementation that uses them.

Especially is enhancing the Error type defined as a submodule of the
providers module for wrapping all the errors returned by any provider to
allow the callers to decide what to do depending on the error "value" and
providing meaningful error messages.