refactor(error): Create an common Error type
Create a common error type which has the most of the error types that
`crate::providers::error` module had for making available not only to
the `crate::providers` but to new submodules that I'll create.

The `crate::providers:error` remains with a specific error type to be
used by the `crate::providers` module and which mostly extends the new
common error type for adding specific error types used by providers
implementations.