refactor(providers/duckdns): Remove `new_` in ctor
Rename the constructor that allows to specify the base URL for not
starting with `new_` prefix because it's more idiomatic without it.

For example the `std::vec::Vec` use `with_` without new for constructing
new instances with some properties and this is like what Duck DNS
provider does with this renamed constructor.