refactor(errors): Generalize external service errors
Duck DNS provider had a variant for returning errors reported by the
provider service, however, the error was generic to any external service
error due to it only specified error common to any external service.

The crate will have other modules that will deal with other external
services, hence, they will likely need to use this "generic" error for
their errors or extends it to add more kinds.

In order of reusing the logic of the external service errors, this
commit has moved the logic to the "crate::error" module but without
adding the variant to its defined `Error` type because not all the
modules may use an external service.

The commit also changed the "Internal" variant to contain a named field
for holding a string for giving a reason why the internal error has
happened.