feat(cli): Implement first commands and flags
Implement the application command-line parameters.

Create the `cli`  module to define all the command-line parameters.

The first version of the application only has one command, `update`. The
command accept a list of providers as positional arguments and an
optional list of finders.

The `cli` module doesn't have any application logic, it only has the
responsibility of parsing and validating the values so the application
can access to them for performing the command logic.