fix(providers/duckdns): Send token & change its type
I forgot to add the token to the request's parameters. This commit
updates the tests to verify that it sends and it adds the token to the
request's parameters list.

I changed the token type because, otherwise, the updater instance is
tied to the token instance, which makes its lifetime weird. On the other
hand, I haven't changed the URL because it's not public and it's a
static string, which only change for testing purposes.