refactor(providers, finders): Create traits
Crate traits for the providers updater functionality and the finder
functionalities because the command-line logic will need to create a
list of all the implementations for exposing via command
parameters/arguments to the user.

Because the functionalities are async and async isn't currently
supported in traits, I had to use the `async_trait` crate.